@@ -1,42 +1,103 @@
 {
-    "name": "flexihash/flexihash",
-    "type": "library",
-    "description": "Flexihash is a small PHP library which implements consistent hashing",
-    "homepage": "https://github.com/pda/flexihash",
+    "name": "esi/consistent-hash",
+    "description": "ConsistentHash is a small library which implements consistent hashing. Modernized fork of flexihash/flexihash.",
     "license": "MIT",
+    "type": "library",
+    "keywords": [
+        "php",
+        "consistent",
+        "hash",
+        "consistent-hash",
+        "consistent-hashing"
+    ],
     "authors": [
+        {
+            "name": "Eric Sizemore",
+            "email": "admin@secondversion.com",
+            "homepage": "https://www.secondversion.com",
+            "role": "Developer"
+        },
         {
             "name": "Paul Annesley",
             "email": "paul@annesley.cc",
-            "homepage": "http://paul.annesley.cc"
+            "homepage": "https://paul.annesley.cc",
+            "role": "flexihash/flexihash Developer"
         },
         {
             "name": "Dom Morgan",
             "email": "dom@d3r.com",
-            "homepage": "https://d3r.com"
+            "homepage": "https://d3r.com",
+            "role": "flexihash/flexihash Developer"
         }
     ],
+    "homepage": "https://github.com/ericsizemore/consistent-hash",
+    "support": {
+        "issues": "https://github.com/ericsizemore/consistent-hash/issues",
+        "forum": "https://github.com/ericsizemore/consistent-hash/discussions",
+        "source": "https://github.com/ericsizemore/consistent-hash/tree/main",
+        "docs": "https://github.com/ericsizemore/consistent-hash",
+        "rss": "https://github.com/ericsizemore/consistent-hash/releases.atom",
+        "security": "https://github.com/ericsizemore/consistent-hash/security/policy"
+    },
     "require": {
-        "php": ">=7.2.0"
+        "php": "^8.2"
     },
     "require-dev": {
-        "phpunit/phpunit": "^8",
-        "squizlabs/php_codesniffer": "3.*",
-        "php-coveralls/php-coveralls": "^2.2",
-        "symfony/config": "^5.1.3",
-        "symfony/console": "^5.1.3",
-        "symfony/filesystem": "^5.1.3",
-        "symfony/stopwatch": "^5.1.3",
-        "symfony/yaml": "^5.1.3"
+        "esi/phpunit-coverage-check": "^2.0",
+        "friendsofphp/php-cs-fixer": "^3.65",
+        "phpbench/phpbench": "^1.3",
+        "phpstan/extension-installer": "^1.4",
+        "phpstan/phpstan": "^2.0",
+        "phpstan/phpstan-deprecation-rules": "^2.0",
+        "phpstan/phpstan-phpunit": "^2.0",
+        "phpstan/phpstan-strict-rules": "^2.0",
+        "phpunit/phpunit": "^11.4",
+        "psalm/plugin-phpunit": "^0.19.0",
+        "vimeo/psalm": "dev-master"
     },
+    "minimum-stability": "dev",
+    "prefer-stable": true,
     "autoload": {
         "psr-4": {
-            "Flexihash\\": "src/"
+            "Esi\\ConsistentHash\\": "src/"
         }
     },
     "autoload-dev": {
         "psr-4": {
-            "Flexihash\\Tests\\": "tests/"
+            "Esi\\ConsistentHash\\Tests\\": "tests/"
         }
+    },
+    "config": {
+        "allow-plugins": {
+            "phpstan/extension-installer": true
+        }
+    },
+    "extra": {
+        "branch-alias": {
+            "dev-main": "1.x-dev"
+        }
+    },
+    "scripts": {
+        "analyze": [
+            "@clear-build",
+            "@composer validate",
+            "@phpunit",
+            "@code-coverage",
+            "@psalm:shepherd",
+            "@psalm:security",
+            "@phpstan"
+        ],
+        "benchmark": "vendor/bin/phpbench run --report=aggregate --iterations=4",
+        "clear-build": "rm -rf ./build/*",
+        "code-coverage": "vendor/bin/coverage-check build/logs/clover.xml 97",
+        "phpcs:check": "vendor/bin/php-cs-fixer fix --config=.php-cs-fixer.dist.php --dry-run --diff",
+        "phpcs:fix": "vendor/bin/php-cs-fixer fix --config=.php-cs-fixer.dist.php",
+        "phpstan": "vendor/bin/phpstan analyse -c phpstan.neon -vvv",
+        "phpunit": "vendor/bin/phpunit",
+        "phpunit:migrate": "@phpunit --migrate-configuration",
+        "psalm": "vendor/bin/psalm -c psalm.xml --no-cache --show-info=true",
+        "psalm:baseline": "@psalm --no-diff --set-baseline=psalm-baseline.xml",
+        "psalm:security": "@psalm --no-diff --taint-analysis",
+        "psalm:shepherd": "@psalm --no-diff --shepherd --stats"
     }
 }
