@@ -1,18 +1,27 @@
 <?php
+
 declare(strict_types=1);
 
-namespace Flexihash\Hasher;
+/**
+ * This file is part of Esi\ConsistentHash.
+ *
+ * (c) Eric Sizemore <admin@secondversion.com>
+ * (c) Paul Annesley <paul@annesley.cc>
+ *
+ * This source file is subject to the MIT license. For the full copyright and
+ * license information, please view the LICENSE file that was distributed with
+ * this source code.
+ */
+
+namespace Esi\ConsistentHash\Hasher;
 
 /**
  * Uses CRC32 to hash a value into a signed 32bit int address space.
  * Under 32bit PHP this (safely) overflows into negatives ints.
- *
- * @author Paul Annesley
- * @license http://www.opensource.org/licenses/mit-license.php
  */
 class Crc32Hasher implements HasherInterface
 {
-    public function hash($string): int
+    public function hash(string $string): int
     {
         return crc32($string);
     }
