@@ -1,13 +1,26 @@
 <?php
+
 declare(strict_types=1);
 
-namespace Flexihash\Hasher;
+/**
+ * This file is part of Esi\ConsistentHash.
+ *
+ * (c) Eric Sizemore <admin@secondversion.com>
+ * (c) Paul Annesley <paul@annesley.cc>
+ *
+ * This source file is subject to the MIT license. For the full copyright and
+ * license information, please view the LICENSE file that was distributed with
+ * this source code.
+ */
+
+namespace Esi\ConsistentHash\Hasher;
+
+use function hexdec;
+use function md5;
+use function substr;
 
 /**
  * Uses MD5 to hash a value into a 32bit int.
- *
- * @author Paul Annesley
- * @license http://www.opensource.org/licenses/mit-license.php
  */
 class Md5Hasher implements HasherInterface
 {
@@ -20,13 +33,9 @@ class Md5Hasher implements HasherInterface
      * The substring is converted to an int since hex strings sometimes get
      * treated as ints if all digits are ints and this results in unexpected
      * sorting order.
-     *
-     * @param  string $string
-     * @return int | float
-     * @author Dom Morgan <dom@d3r.com>
      */
-    public function hash($string)
+    public function hash(string $string): int
     {
-        return hexdec(substr(md5($string), 0, 8));
+        return (int) hexdec(substr(md5($string), 0, 8));
     }
 }
